pyproject.toml
[tool.poetry]
name = "pgpythonshop"
version = "0.1.0"
description = ""
authors = ["Artemowkin <artyom.groshev2017@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
